\name{eenvlp-package}
\alias{eenvlp-package}
\docType{package}
\title{
Enhanced response envelope
}
\description{
This package provides the implementation of the enhanced response envelope. It gives you the enhanced envelope estimator for multivariate linear regression problem. \cr
}

\details{
The main function of the package \code{\link{eenvlp}} is \code{eenvlp}. \cr
}

\keyword{envelope}
\keyword{multivariate response}
\keyword{dimension reduction}
\keyword{package}

