\name{eenvlp}
\alias{eenvlp}
\title{Fit the enhanced response envelope model}
\description{
 Fit the enhanced response envelope model in multivariate linear regression with dimension u. 
}
\usage{
eenvlp(X, Y, u, lamb, init = NULL, X.scale = TRUE)
}
\arguments{
  \item{X}{Predictors. An n by p matrix, p is the number of predictors. The predictors can be univariate or multivariate, discrete or continuous.}
  \item{Y}{Multivariate responses. An n by r matrix, r is the number of responses and n is number of observations. The responses must be continuous variables.}
  \item{u}{Dimension of the envelope. An integer between 0 and r.}
  \item{lamb}{A regularization parameter. A positive value.}    
  \item{init}{The user-specified value of Gamma for the envelope subspace. An r by u matrix. The default is the one generated by function envMU.}
  \item{X.scale}{A boolean. If TRUE (the default), then the columns of training X are divided by their standard deviations. If False, no scaling is done.}
}
\details{
 This function fits the envelope model to the responses and predictors, \deqn{
 Y = \mu + \Gamma\eta X+\varepsilon, \Sigma=\Gamma\Omega\Gamma'+\Gamma_{0}\Omega_{0}\Gamma'_{0}
 }
 using the penalized maximum likelihood estimation.  When the dimension of the envelope is between 1 and r-1, the starting value and blockwise coordinate descent algorithm in Cook et al. (2016) is implemented.  When the dimension is r, then the envelope model degenerates to the standard multivariate linear regression.  When the dimension is 0, it means that X and Y are uncorrelated, and the fitting is different.
}
\value{The output is a list that contains the following components:
\item{beta}{The envelope estimator of the regression coefficients.} 
\item{Sigma}{The envelope estimator of the error covariance matrix.}
\item{Gamma}{An orthonormal basis of the envelope subspace.}
\item{Gamma0}{An orthonormal basis of the complement of the envelope subspace.}
\item{eta}{The coordinates of beta with respect to Gamma.}
\item{Omega}{The coordinates of Sigma with respect to Gamma.}
\item{Omega0}{The coordinates of Sigma with respect to Gamma0.}
\item{mu}{The estimated intercept.}
\item{n}{The number of observations in the data.}
\item{X.center}{The numeric centering used.}
\item{X.scale}{The numeric scaling used. If not used, NULL.}
}
\references{
Cook, R. D., Forzani, L. and Su, Z. (2016) A Note on Fast Envelope Estimation. Journal of Multivariate Analysis. 150, 42-54.
}

\examples{
data(cereal)
m <- eenvlp(cereal$x, cereal$y, u=1, lamb=1e-07)
m$Gamma
}
